pyproject.toml
[project]
name = "drf-p37"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "automium-web>=0.1.1",
    "django>=6.0.4",
    "django-filter>=25.2",
    "djangorestframework>=3.17.1",
    "drf-spectacular>=0.29.0",
    "pillow>=12.2.0",
    "psycopg2-binary>=2.9.11",
    "python-dotenv>=1.2.2",
    "redis>=7.4.0",
]
